Reduce board size

+++ engine/src/material.hpp
@@ -118,7 +118,7 @@ namespace wisdom
         int my_score[Num_Players] {};
 
         // Count of pieces on either side:
-        array<array<int, Num_Piece_Types>, Num_Players> my_piece_count {};
+        array<array<int8_t, Num_Piece_Types>, Num_Players> my_piece_count {};
     };
 }
